pyproject.toml
[tool.poetry]
name = "airflowx"
version = "0.1.0"
description = ""
authors = ["amArbaoui <adel.arbaoui@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "~1.4"

[tool.poetry.group.dev.dependencies]
testcontainers = "^4.3.3"
pytest = "^8.1.1"
black = "^24.3.0"
mypy = "^1.9.0"
ruff = "^0.3.5"
psycopg2-binary = "^2.9.9"
pymysql = "^1.1.0"
pytest-cov = "^5.0.0"
apache-airflow-providers-postgres = "^5.10.2"
pendulum = "2.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
